add useful structs and helper functions to generate the key parts of the envoy config
